requirements.txt
ib_insync
pandas
numpy
python-dotenv
requests
pytz
